requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
orjson>=3.0.0
python-dotenv>=1.0.0
GitPython>=3.1.0